# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Core Topology Generator Web API
  version: "2026-03-05"
  description: >-
    OpenAPI specification for the CORE topology generator web backend. The API
    combines JSON endpoints used by the web UI with a set of HTML form
    submissions and download helpers. Most routes require an authenticated
    session established via `/login` and maintained with the Flask session
    cookie.
servers:
  - url: http://localhost:9090
    description: Default development server
  - url: https://localhost
    description: Reverse-proxied TLS endpoint (see README for nginx helper)
...
security:
  - sessionCookie: []
paths:
...
  /api/scenario/latest_xml:
    get:
      tags:
        - Scenario Editor
      summary: Get the latest saved XML path for a scenario
      parameters:
        - name: scenario
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Latest XML path for the scenario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestScenarioXmlResponse'
        '400':
          description: Missing scenario name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No XML found for the scenario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/scenario/latest_state:
    get:
      tags:
        - Scenario Editor
      summary: Parse the latest saved scenario state from XML
      description: >-
        Returns the editor-shaped scenario JSON plus the top-level core configuration for the
        requested scenario. This is useful for round-tripping saved XML back into structured state.
      parameters:
        - name: scenario
          in: query
          required: true
          schema:
            type: string
        - name: xml_path
          in: query
          required: false
          schema:
            type: string
          description: Optional explicit XML path to parse instead of the latest saved XML.
      responses:
        '200':
          description: Parsed scenario state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestScenarioStateResponse'
        '400':
          description: Missing scenario name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: XML not found or scenario missing from XML.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/open_scripts:
    get:
      tags:
        - Scripts
      summary: List generated traffic or segmentation scripts
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [traffic, segmentation]
            default: traffic
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [runtime, preview]
            default: runtime
      responses:
        '200':
          description: Script listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenScriptsResponse'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested directory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/open_script_file:
    get:
      tags:
        - Scripts
      summary: Read the contents of a generated script file
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [traffic, segmentation]
            default: traffic
        - name: scope
          in: query
          schema:
            type: string
            enum: [runtime, preview]
            default: runtime
        - name: file
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Script preview (first 8KB).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenScriptFileResponse'
        '400':
          description: Invalid parameters or filename.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Script directory or file not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/download_scripts:
    get:
      tags:
        - Scripts
      summary: Download generated traffic or segmentation scripts as a zip archive
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [traffic, segmentation]
            default: traffic
        - name: scope
          in: query
          schema:
            type: string
            enum: [runtime, preview]
            default: runtime
      responses:
        '200':
          description: Zip archive of generated scripts.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Requested script directory not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie issued after authenticating with `/login`.
  schemas:
...
    LatestScenarioXmlResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
        scenario:
          type: string
        xml_path:
          type: string
        error:
          type: string
    LatestScenarioStateResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
        scenario:
          type: string
        scenario_norm:
          type: string
        xml_path:
          type: string
        xml_mtime:
          type: number
        scenario_state:
          type: object
          additionalProperties: true
        core:
          type: object
          nullable: true
          additionalProperties: true
        error:
          type: string
...
    ScriptFileSummary:
      type: object
      properties:
        file:
          type: string
        size:
          type: integer
    OpenScriptsResponse:
      type: object
      properties:
        ok:
          type: boolean
        kind:
          type: string
          enum: [traffic, segmentation]
        path:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/ScriptFileSummary'
        error:
          type: string
      required:
        - ok
    OpenScriptFileResponse:
      type: object
      properties:
        ok:
          type: boolean
        file:
          type: string
        path:
          type: string
        content:
          type: string
        truncated:
          type: boolean
        error:
          type: string
      required:
        - ok
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          nullable: true
        error:
          type: string
...